refactor: Update configuration and processing for geometric feature extraction; remove unused point cloud normal calculation script

+++ processing/spherical_back_projection.py
@@ -4,7 +4,7 @@ Run this file with `python run_3d_to_2d_pipeline.py` or `python -m processing.sp
 """
 
 import numpy as np
-from tools.illustrate_spherical_projection import (generate_lidar_ball, 
+from tools.illustrate_spherical_projection import (generate_geometric_lidar_ball, 
                                              visualize_and_save_point_cloud, 
                                              create_colored_cube_points)
 from PIL import Image
@@ -13,8 +13,10 @@ from tools.config_loader import CONFIG
 from pathlib import Path
 from tools.preprocess_point_cloud import map_angle_to_pixel
 from plyfile import PlyData, PlyElement
-from processing.spherical_projection import load_image_cube_and_meta
+# from processing.spherical_projection import load_image_cube_and_meta
+from tools.spherical_projection_helper import load_image_cube_and_meta
 from numpy.typing import NDArray
+from math import ceil
 
 def image_preprocess(image, img_size):
     """
@@ -32,10 +34,14 @@ def image_preprocess(image, img_size):
 
     # Resize to (271, 720) if needed (zenith: 0–135 at 0.5° → 271 rows)
     if image.dtype == np.float64:
-        image_uint8 = (np.clip(image, 0, 1) * 255).astype(np.uint8)
+        if image.max() > 1.0 or image.min() < 0.0:
+            image = (image - image.min()) / (image.max() - image.min())
+        image_uint8 = (image * 255).astype(np.uint8)
         img_resized = Image.fromarray(image_uint8).resize((img_size[1], img_size[0]))
-    else:
+    elif image.dtype == np.uint8:
         img_resized = Image.fromarray(image).resize((img_size[1], img_size[0]), resample=Image.BILINEAR)
+    else:
+        raise ValueError(f"Unsupported image dtype: {image.dtype}. Expected uint8 or float64.")
 
     image = np.array(img_resized)
 
@@ -87,7 +93,8 @@ def get_a_colorized_ball_from_img(rgb_img: NDArray,
                                   key_str: str,
                                   zenith_range: tuple = (0, 135), 
                                   save_dir: Path = None,
-                                  visualize: bool = False):
+                                  visualize: bool = False,
+                                  inverse_zenith: bool = False):
     """
     Get a colorized ball from an image using spherical projection.
     Args:
@@ -96,10 +103,10 @@ def get_a_colorized_ball_from_img(rgb_img: NDArray,
         zenith_range (tuple): Zenith range for the ball.
         save_dir (Path): Directory to save the output point cloud.
     """
-
+    df_ball = generate_geometric_lidar_ball(radius=10.0, zenith_range=zenith_range)
     df_colored = back_project_color_to_ball(df_ball, rgb_img, 
                                             zenith_range=zenith_range,
-                                            inverse_zenith=False)
+                                            inverse_zenith=inverse_zenith)
 
     df_cube = create_colored_cube_points(center=[0, 0, 0], size=0.5, samples_per_face=50)
     df_combined = pd.concat([df_colored, df_cube], ignore_index=True)
@@ -110,10 +117,11 @@ def get_a_colorized_ball_from_img(rgb_img: NDArray,
                                 output_prefix=key_str)
 
 
-def attach_image_colors_to_pcd(rgb_image, pcd_out_dir, 
+def attach_image_colors_to_pcd(rgb_image, pcd_out_dir, input_pcd_key: str,
                                canvas_size: tuple[int, int],
                                 angular_res: tuple[int, int],
-                                channel_names=['r', 'g', 'b']):
+                                channel_names=['r', 'g', 'b'],
+                                delete_intermediate: bool = False):
     """
     Attach RGB or arbitrary 3-channel color values from a 2D image to a point cloud based on pixel mapping.
 
@@ -127,18 +135,18 @@ def attach_image_colors_to_pcd(rgb_image, pcd_out_dir,
 
     point_cloud_file = next(pcd_out_dir.glob(f"*_color*"), None)
     if point_cloud_file is None:
-        point_cloud_file = next(pcd_out_dir.glob(f"*_filtered_normaled*"), None)
-    
+        point_cloud_file = next(pcd_out_dir.glob(f"*{input_pcd_key}*"), None)
 
     pc_df = pd.read_csv(point_cloud_file, sep=',')
 
     if "_color" in point_cloud_file.stem:
         output_file = pcd_out_dir / (point_cloud_file.stem + '.csv')
     else:
-        file_str = str(point_cloud_file.stem).split("_filtered_normaled")[0]
+        file_str = str(point_cloud_file.stem).split(input_pcd_key)[0]
         output_file = pcd_out_dir / (f"{file_str}_color" + '.csv')
-        point_cloud_file.unlink()
-        print(f"Deleted intermediate file: {point_cloud_file}")
+        if delete_intermediate:
+            point_cloud_file.unlink()
+            print(f"Deleted intermediate file: {point_cloud_file}")
         
     assert rgb_image.ndim == 3 and rgb_image.shape[2] == 3, "Image must have 3 channels"
     assert len(channel_names) == 3, "Exactly 3 channel names must be provided"
@@ -173,50 +181,69 @@ def attach_image_colors_to_pcd(rgb_image, pcd_out_dir,
 
 
 if __name__ == "__main__":
-    v_fov = CONFIG['global']['v_fov']
-    h_fov = CONFIG['global']['h_fov']
-    canvas_width = int(h_fov[1] / CONFIG['global']['h_ang_res_deg'])
-    canvas_height = int(v_fov[1] / CONFIG['global']['v_ang_res_deg'])
+
+    # Load configuration
+    global_params = CONFIG['global']
+    v_fov = global_params['v_fov']
+    h_fov = global_params['h_fov']
+    angular_res = (global_params['v_ang_res_deg'], global_params['h_ang_res_deg'])
+    canvas_width = int(ceil((h_fov[1]-h_fov[0]) / angular_res[1]))
+    canvas_height = int(ceil((v_fov[1]-v_fov[0]) / angular_res[0]))
     canvas_size = (canvas_height, canvas_width)
-    angular_res = (CONFIG['global']['v_ang_res_deg'], CONFIG['global']['h_ang_res_deg'])
-    zenith_range = CONFIG['global']['v_fov']  # Zenith range for the CBL V2.0
-    df_ball = generate_lidar_ball(radius=10.0, zenith_range=zenith_range)
-    
-    out_dir_ls = CONFIG['global']['output_dir_ls']
+    print(f"Canvas Size: {canvas_size}, Angular Resolution: {angular_res}")
+    zenith_range = global_params['v_fov']
+    delete_intermediate_file = global_params['delete_intermediate_file']
+    out_dir_ls = global_params['output_dir_ls']
+    out_signature_str = CONFIG['calc_geom_feature']['out_signature_str']
+    inverse_zenith = CONFIG['calc_geom_feature']['flip_mangrove']
+
     for output_dir in out_dir_ls:
         input_file_stem = output_dir.parent.name
         print(f'#######Processing {input_file_stem}...########')
 
         image_dir = output_dir / 'img'
         pcd_out_dir = output_dir / 'pcd'
-
         
         key_str = input_file_stem.split('_')[0] + '_' + input_file_stem.split('_')[-1]
         image_cube_path = image_dir / f'{key_str}_image_cube.npy'
         image_cube, metadata = load_image_cube_and_meta(image_cube_path)
 
         channel_names = metadata['channel_names']
-        
-        if channel_names[0]== "True-R":
-            group_indices = [
-                            [6, 7, 8], 
-                            [9, 10, 11], 
-                            [12, 13, 14]]
-        elif channel_names[0]== "Intensity Map (raw)":
-            group_indices = [[3, 4, 5], 
-                            [6, 7, 8], 
-                            [9, 10, 11], 
-                            ]
-        else:
-            raise ValueError(f"Weird channel name: {channel_names[0]}, checkout previous script.")
-        channel_name_groups = [[channel_names[i] for i in group] for group in group_indices]
-        rgb_groups = [image_cube[:,:, group_index] for group_index in group_indices]
-        get_a_colorized_ball_from_img(image_cube[:,:, [0,1,2]], 
-                                    key_str,
-                                    zenith_range=zenith_range, 
-                                    save_dir=pcd_out_dir)
-        for channel_name_group, rgb_image in zip(channel_name_groups, rgb_groups):
-            attach_image_colors_to_pcd(rgb_image, pcd_out_dir, 
+        paint_pcd_color_groups = [
+            ['intensity_adjusted', 'z_adjusted', 'range_adjusted'],
+            ['curvature', 'anisotropy', 'planarity'],
+            ['Pseudo-Rn', 'Pseudo-Gn', 'Pseudo-Bn'],
+        ]
+        # check if all the names in the paint_pcd_color_groups are in channel_names
+        for group in paint_pcd_color_groups:
+            for name in group:
+                if name not in channel_names:
+                    raise ValueError(f"Channel name {name} not found in image cube metadata. Available channels: {channel_names}")
+
+        # Paint the point cloud with different color groups
+        for channel_name_group in paint_pcd_color_groups:
+            rgb_image = image_cube[:, :, [channel_names.index(name) for name in channel_name_group]]
+            attach_image_colors_to_pcd(rgb_image, pcd_out_dir, out_signature_str,
                                        channel_names=channel_name_group, 
                                        canvas_size=canvas_size,
-                                       angular_res=angular_res)
+                                       angular_res=angular_res,
+                                       delete_intermediate=delete_intermediate_file)
+
+        # Generate colorized balls from certain color groups
+        paint_ball_color_groups = [
+            ['intensity_adjusted', 'z_adjusted', 'range_adjusted'],
+            ['Pseudo-Rn', 'Pseudo-Gn', 'Pseudo-Bn'],
+            ['PCA1', 'PCA2', 'PCA3']
+        ]
+        if 'True-R' in channel_names:
+            paint_ball_color_groups.append(['True-R', 'True-G', 'True-B'])
+        
+        for color_group in paint_ball_color_groups:
+            key_str = f"{input_file_stem}_{'_'.join(color_group)}"
+            rgb_image = image_cube[:, :, [channel_names.index(name) for name in color_group]]
+            get_a_colorized_ball_from_img(rgb_image,
+                                    key_str,
+                                    zenith_range=zenith_range, 
+                                    save_dir=pcd_out_dir,
+                                    inverse_zenith=inverse_zenith)
+
